# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sede",
        "description": "Listado de las URL de la Api Sede",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/sedes": {
            "get": {
                "tags": [
                    "Sede"
                ],
                "summary": "Listar todas las sedes",
                "operationId": "2c1ea09725b25fcd2420322790f36162",
                "responses": {
                    "200": {
                        "description": "Vista con listado de sedes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sede"
                ],
                "summary": "Guardar una nueva sede",
                "operationId": "2387d1d5c64447fd136d89893bab3201",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion",
                                    "tipo",
                                    "capacidad_estudiantes",
                                    "carreras_ofrecidas"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sede Norte"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Calle Falsa 123"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "instituto"
                                    },
                                    "capacidad_estudiantes": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "carreras_ofrecidas": {
                                        "type": "string",
                                        "example": "Ingeniería, Medicina"
                                    },
                                    "encargado_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirección al listado de sedes"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
        "/sedes/{id}": {
            "get": {
                "tags": [
                    "Sede"
                ],
                "summary": "Mostrar detalle de una sede",
                "operationId": "7c18207c51ff825ed001960ae413741e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vista con detalle de la sede"
                    },
                    "404": {
                        "description": "Sede no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sede"
                ],
                "summary": "Actualizar sede",
                "operationId": "7ede1848e7f1f0cb5a4396128f6916ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion",
                                    "tipo",
                                    "capacidad_estudiantes",
                                    "carreras_ofrecidas"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sede Sur"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Siempre Viva 742"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "universidad"
                                    },
                                    "capacidad_estudiantes": {
                                        "type": "integer",
                                        "example": 300
                                    },
                                    "carreras_ofrecidas": {
                                        "type": "string",
                                        "example": "Contabilidad, Derecho"
                                    },
                                    "encargado_id": {
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirección al listado de sedes"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sede"
                ],
                "summary": "Eliminar sede",
                "operationId": "da7fc323aed54526c521d1fa4573a618",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirección al listado de sedes con mensaje de éxito"
                    },
                    "404": {
                        "description": "Sede no encontrada"
                    }
                }
            }
        },
        "/sedes/{id}/edit": {
            "get": {
                "tags": [
                    "Sede"
                ],
                "summary": "Formulario de edición de una sede",
                "operationId": "8eebbc8a1f3ea1161381a57852b8474e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vista con formulario de edición de sede"
                    },
                    "404": {
                        "description": "Sede no encontrada"
                    }
                }
            }
        },
...
    },
...
}